Revert "Added code for returning evidence instead of 1 for winning class, but not sure it's working"
This reverts commit 98d6e3eaa86dd0572f967c5c79b2cedae9561545.

+++ PairwiseTrainingState.java
@@ -23,7 +23,7 @@ public class PairwiseTrainingState {
 	/**
 	 * Testing holders for models
 	 */
-	protected double[] votes;
+	protected int[] votes;
 	protected double[] evidence;
 	
 	public PairwiseTrainingState(){
@@ -53,11 +53,11 @@ public class PairwiseTrainingState {
 		this.labelNumber = labelNumber;
 	}
 
-	public double[] getVotes() {
+	public int[] getVotes() {
 		return votes;
 	}
 
-	public void setVotes(double[] votes) {
+	public void setVotes(int[] votes) {
 		this.votes = votes;
 	}
